requirements.txt
asgiref==3.3.4
Django==3.2.1
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
<<<<<<< HEAD
=======
psycopg2>=2.8.6,<2.8.7
>>>>>>> 9c9c0a9aa7616dd13695887579ecfc9732953972
